gh-79315: Add Include/cpython/modsupport.h header
